Data.Enumerate.Domain

Description

e.g. generate a boilerplate HasTrie instance for a Context type that has both sum types and product types

data Context
 = GlobalContext | EmacsContext BufferName BufferMode BufferContents | OtherContext String
 deriving (Show,Generic) 
 -- Strings are infinite, so Context is too, that's okay 

instance HasDomain Context 

type BufferName = String
type BufferMode = String
type BufferContents = String

instance HasTrie Context where
 type a :->: b = ContextTrie (TrieGeneric a b)
 trie = trieGeneric
 untrie = untrieGeneric

the conversions (between a type a and its representation (Rep a x)) make the generic instance slower than a manual instance.

but, if the function you're memoizing is costly enough, and the datatype it consumes is messy enough, the four-line HasTrie instance (and a one line call to memo) can save both runtime and developer time.