Data.Enumerate.Extra

Description

Documentation

failed :: MonadThrow m => String -> m a

failed = throwM . userError

maybe2throw :: (a -> Maybe b) -> forall m. MonadThrow m => a -> m b

generalize a function that fails with Nothing.

list2throw :: (a -> [b]) -> forall m. MonadThrow m => a -> m b

generalize a function that fails with [].

either2throw :: (a -> Either SomeException b) -> forall m. MonadThrow m => a -> m b

generalize a function that fails with Left.

totalizeFunction :: (NFData b, MonadThrow m) => (a -> b) -> a -> m b

makes an *unsafely*-partial function (i.e. a function that throws exceptions or that has inexhaustive pattern matching) into a *safely*-partial function (i.e. that explicitly returns in a monad that supports failure).

defaultPartialityHandlers :: MonadThrow m => [Handler (m a)]

handles the following exceptions:

ArithException

ArrayException

ErrorCall

PatternMatchFail

spoonWith :: (NFData a, MonadThrow m) => [Handler (m a)] -> a -> m a

Evaluate a value to normal form and throwM any exceptions are thrown during evaluation. For any error-free value, spoon = Just.

taken from the spoon package.

showsPrecWith :: (Show a, Show b) => String -> (a -> b) -> Int -> a -> ShowS

the eliminator as a function and the introducer as a string

helper for declaring Show instances of datatypes without visible constructors (like Map which is shown as an list).

int2natural :: Int -> Natural

powerSet :: Ord a => Set a -> Set (Set a)

the power set of a set of values.

>>> (powerset2matrix . powerSet . Set.fromList) [1..3]
[[],[1],[2],[3],[1,2],[1,3],[2,3],[1,2,3]]

dropEach :: Ord a => Set a -> Set (Set a)

>>> (powerset2matrix . dropEach . Set.fromList) [1..3]
[[1,2],[1,3],[2,3]]

powerset2matrix :: Set (Set a) -> [[a]]

convert a power set to an isomorphic matrix, sorting the entries.

(for doctest)