Enumerate.Types

Description

enumerate all values in a finite type.

see the Enumerable class for documentation.

see Enumerate.Example for examples.

can also help automatically derive QuickCheck instances:

newtype ValidString = ValidString String
 deriving (Show)
validStrings :: [String]
makeValidString :: String -> Maybe ValidString
makeValidString s = if s member validStrings then Just (ValidString s) else Nothing
instance Enumerable ValidString where enumerated = ValidString $ validStrings ... -- manually (since normal String's are infinite)
instance Arbitrary ValidString where arbitrary = elements enumerated

data ValidName = ValidName ValidString ValidString | CoolValidName [ValidString]
 deriving (Show,Generic)
instance Enumerable ValidName -- automatically

instance Arbitrary ValidName where arbitrary = elements enumerated

Provides instances for all base types (whenever possible):

under Data. / Control. / System. / Text., and even GHC.

even non-Enums

except when too large (like Int) (see Enumerate.Large)

background on Generics:

GHC.Generics

also provides instances for:

sets

vinyl records

related packages:

enumerable. no Generic instance.

universe no Generic instance.

SafeEnum only Enums

emgm. allows infinite lists (by convention). too heavyweight.

testing-feat. too heavyweight (testing framework).

smallcheck too heavyweight (testing framework). Series enumerates up to some depth and can enumerated infinitely-inhabited types.

quickcheck too heavyweight (testing framework, randomness unnecessary).

modular integers

class Enumerable a where

enumerate the set of all values in a (finitely enumerable) type. enumerates depth first.

generalizes Enums to any finite/discrete type. an Enumerable is either:

an Enum

a product of Enumerables

a sum of Enumerables

can be implemented automatically via its Generic instance.

laws:

finite:

cardinality /= _|_

consistent:

cardinality _ = length enumerated

so you can index the enumerated with a nonnegative index below the cardinality.

distinct:

(Eq a) => nub enumerated == enumerated

complete:

x `elem' enumerated

coincides with Bounded Enums:

(Enum a, Bounded a) => enumerated == boundedEnumerated

(Enum a) => enumerated == enumEnumerated

(Bounded constraint elided for convenience, but relevant.)

("inputs" a type, outputs a list of values).

Minimal complete definition

Nothing

Methods

enumerated :: [a]

cardinality :: proxy a -> Natural

newtype WrappedBoundedEnum a

wrap any (Bounded a, Enum a) to be a Enumerable via boundedEnumerated.

(avoids OverlappingInstances).

Constructors

WrappedBoundedEnum

Fields

unwrapBoundedEnum :: a

class GEnumerable f where

"Generic Enumerable", lifted to unary type constructors.

Methods

genumerated :: [f x]

gcardinality :: proxy f -> Natural

boundedEnumerated :: (Bounded a, Enum a) => [a]

for non-Generic Bounded Enums:

instance Enumerable _ where
 enumerated = boundedEnumerated
 cardinality = boundedCardinality

boundedCardinality :: forall proxy a. (Bounded a, Enum a) => proxy a -> Natural

for non-Generic Bounded Enums.

Assuming Bounded is correct, safely stop the enumeration (and know where to start).

behavior may be undefined when the cardinality of a is larger than the cardinality of Int. this should be okay, as Int is at least as big as Int64, which is at least as big as all the monomorphic types in base that instantiate Bounded. you can double-check with:

>>> boundedCardinality (const(undefined::Int))   -- platform specific
18446744073709551616

-- i.e. 1 + 9223372036854775807 - (-9223372036854775808)

works with non-zero-based Enum instances, like Int64 or a custom toEnum/fromEnum. assumes the enumeration's numbering is contiguous, e.g. if fromEnum 0 and fromEnum 2 both exist, then fromEnum 1 should exist too.

enumEnumerated :: Enum a => [a]

for non-Generic Enums:

instance Enumerable ... where
 enumerated = enumEnumerated

the enum should still be bounded.

indexedEnumerated :: (Bounded a, Ix a) => [a]

for non-Generic Bounded Indexed (Ix) types:

instance Enumerable _ where
 enumerated = indexedEnumerated
 cardinality = indexedCardinality

indexedCardinality :: forall proxy a. (Bounded a, Ix a) => proxy a -> Natural

for non-Generic Bounded Indexed (Ix) types.

enumerateBelow :: forall a. Enumerable a => Natural -> Either Natural [a]

enumerate only when the cardinality is small enough. returns the cardinality when too large.

>>> enumerateBelow 2 :: Either Natural [Bool]
Left 2

>>> enumerateBelow 100 :: Either Natural [Bool]
Right [False,True]

useful when you've established that traversing a list below some length and consuming its values is reasonable for your application. e.g. after benchmarking, you think you can process a billion entries within a minute.

enumerateTimeout :: (Enumerable a, NFData a) => Int -> IO (Maybe [a])

enumerate only when completely evaluating the list doesn't timeout (before the given number of microseconds).

>>> enumerateTimeout (2 * 10^6) :: IO (Maybe [Bool])  -- two seconds
Just [False,True]